socialn etwork/social_network/requirements.txt
amqp==5.2.0
asgiref==3.7.2
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Django==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.0
django-filter==24.3
django-timezone-field==7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
firebase-admin==6.5.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
grpcio==1.62.1
kombu==5.3.7
lxml==5.2.1
mysqlclient==2.2.4
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2023.3.post1
redis==5.0.7
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.2.1
whitenoise==6.6.0
